Fix handling of control characters in EditableText type
For whatever reason, the byte_index() function was taking into account a
grapheme's width when determining the byte index of a "character". Given
that control characters have a width of zero, this necessarily leads to
a wrong result once such characters are involved.
This change fixes the problem by adjusting the byte_index() function to
ignore the width and just treat each grapheme equally.