Rename char_index() function to cursor_index()
As it turns out, the char_index() is admittedly a bit of a misnomer:
because it takes into account the grapheme's width, it does not really
report a character index in a strict sense. Rather, it is about the
cursor, which does not care about entities such a control characters.
This change renames the function to cursor_index() in an attempt to make
this distinction more obvious and prevent misuse. One such case of
misuse is in the Text::len() method, which should not ignore control
characters. With this change we replace the usage of this function in
there with a mere count of the graphemes.