Fix clipping logic
As it turns out our clipping logic is subtly broken in the presence of
multi-byte characters. Because we unconditionally cut off the input
string once we exceeded a number of bytes, we may inadvertently violate
character boundaries. That in turn can result in a panic being reported.
This change fixes the problem by introducing a proper clipping routine
that honors graphemes and will only cut off entire graphemes instead of
individual bytes. It will also take into account the actual display
width of the string as opposed to merely counting bytes.