Generalize and rename sanitize_offset() function
The sanitize_offset() function is arguably somewhat hard to wrap one's
head around, owing in part to its somewhat nondescript name but also the
perhaps non-intuitive functionality that it provides.
Thinking about said functionality in terms of a window into a range of
values and a cursor pointing to the value that is most certainly meant
to be included hopefully makes it easier to grasp. To that end, this
change renames the function to window_start(). It also generalizes it in
anticipation of future changes that may want to invoke it on non-usize
values.