Remove Deref<Target = str> impl for Text
Currently a Text object transparently derefs into a &str. That is
arguably confusing in and of itself, but it gets compounded by the fact
that both &str and Text have a len() method -- which have different
semantics.
This change fixes both issues. First, it removes the Deref impl and then
it renames the Text::len() method to char_count(), which more accurately
reflects its purpose.