Rework cursor and character editing logic
We have struggled royally getting the cursor and character editing logic
correct in the presence of multi-byte graphemes as well as multi-column
Unicode characters. One of the main reasons for this struggle was that
we are straddling (but haven't clearly differentiated) two different
domains: one is the domain of characters (Unicode graphemes), which can
be removed and added. The other being the cursor being displayed and
representing the point at which the next edit operation is taking place.
Because the latter take into account the display width of a
character/string, but the former work solely in the realm of bytes, it
is easy to wrongly intermix the two domains. That can and does lead to
subtle bugs. What makes matters worse is that at least the cursor
display logic is hard (impossible?) to test programmatically (but
relatively easy to do "by eye" given the right test data).
This change attempts to rectify this situation once and for all. It
introduces the Cursor type that represents a cursor as it is displayed.
Objects of this type mostly originate from a single function and cannot
easily be conjured up accidentally by a user. They interact with the
Width type, which is also introduced, and which represents the width of
a string/character. A Width object, while user-constructible, is
typically mostly created via the DisplayWidth trait (implemented for
&str and char). Creating such objects via a trait is essential to lower
the barrier to usage throughout the program and -- so is the hope --
will make it easier to permeate correct Unicode logic into more parts of
the program that still are not strictly speaking correct for general
Unicode inputs. In short, with the introduction of this infrastructure
we make it extremely unlikely that the two domains are accidentally
mixed up.