Implement logic for wrapping long lines at a particular width
So far all of our input logic has been a single line. As such, there has
not been any need to perform wrapping. Instead, if something did not fit
we just cut it off and then we had the means to scroll content. With
upcoming changes this state of affairs will change and we will properly
support multi-line text in some cases. To display that in a reasonable
way on a restricted terminal, it is best to break up overly long lines
and let them flow over into the next line.
This change introduces the necessary logic for detecting and breaking up
such a long line so at a word boundary.