Add "details" to Task type
So far our tasks are mainly comprised of a summary along with a set of
tags. This covers a wide variety of use cases, but we have had more and
more instances where additional storage for notes and details would be
helpful.
To that end, this change adds another string to Task objects, which may
hold arbitrary "details". Contrary to the summary, this string is may
hold a basically arbitrary number of lines. In the iCal backend this
field maps to a TODO's description.