Correctly handle multi-line text in EditableText
The EditableText type (and, more generally, the Unicode logic
surrounding it) does not currently support newline symbols properly.
That is because they are control characters and as are defined to have a
width of zero. However, that goes against our vision of how to treat
them in the program.
With this change we treat newlines as having a width of one, which means
that they can be navigated to with a cursor, for example.