Add multi-line support to InputText type
Our InputText type currently only supports single line input (at least
when not on compiled with readline support enabled). Our task details
are meant to be a multi-line string.
To make editing them as such possible, this change enhances the
InputText type with optional multi-line support. In a nutshell, we add
the means for handling the Alt-Return key sequence and insert a carriage
return into the text when encountering it. Support for multi-line
semantics has to be explicitly enabled via the newly introduced
InputTextBuilder.