Hook up DetailDialog
This change hooks up the DetailDialog widget to the rest of the program.
That is, it puts in place the necessary plumbing for creating the widget
and adjusts the terminal based renderer to render it, if active.
The dialog can be opened by using the return key on the selected task in
the task list box. When the program is compiled with readline support,
said dialog will support readline bindings. Note, however, that the
input is treated as a single line from libreadline's point of view,
because the library is designed for editing a single line.