Stop using implicit dependency features
Our optional dependencies were implicitly defining features of the same
name, as that is how Cargo opted to do things. The behavior is weird and
confusing and best avoided. With version 1.60, we can opt out by making
actual features depend on said optional dependencies via dep:<name> [0].
With this change we do just that and are able to remove some package
renaming kludge that we had to put in place because of this behavior.

[0] https://doc.rust-lang.org/cargo/reference/features.html#optional-dependencies