Optimize "line filling" logic
Our dialog widgets use various fill_*_line() methods that fill a part or
the rest of a line with a certain color. These methods are excessively
inefficient in terms of performance, because they fill each and every
cell using a dedicated write call that moves the cursor and sets the
foreground and background color. While this has been acceptable practice
so far, it will bite us down the line.
This change optimizes this logic, introducing first-class support for
line filling to the ClippingWriter type and using that to replace all
existing fill_*_line() methods. In the process we also fix a bug where
we used the fill_tag_dialog_line() helper for rendering parts of the
detail dialog, which could have resulted in the wrong color being used.