Refactor and improve TaskListBox rendering logic
In terms of how we go about rendering the TaskListBox widget, the logic
is a bit of a snowflake compared to that of all the other widgets.
Specifically, all widgets basically take a line based approach, where
each line is filled as appropriate. For the TaskListBox, we rather focus
on the tasks themselves. Not only is that an unnecessary divergence, the
current logic also does not correctly fill the space between two items
or even just the unused remainder of a line. That is not currently a
problem, because we redraw everything anyway, and so everything looks
correct because the parent widget makes sure that it is, but if we
switch to a model in which only changed widgets are re-rendered, this
would cause problems.
This change refactors the TaskListBox rendering logic to be line based
as well and brings everything closer to what we do for other widgets. It
also makes sure that we render all lines and fill each one completely,
including otherwise unused space.