Minimize usage of MessageExt::maybe_update method
We use the MessageExt::maybe_update() helper method in quite a few
locations, including where it arguably is unnecessary, because, at least
with somewhat recent versions of Rust, we can achieve shorter code by
using the bool::then_some() combinator instead.
With this change we adjust a bunch of occurrences of the former with the
latter construct.