Mitigate potential clip underflow in clip() function
It is conceivable that a caller provides a x value to the clip()
function that actually exceeds the width of the current bounding box.
Given that this function works with unsigned values, that would result
in an underflow, which could trigger a panic.
While not exactly expected, this input is also not really a bug. As
such, this change hardens the function against such panics by using
saturating subtraction instead.