Adjust MessageExt::maybe_update to require Option<Message>
So far our MessageExt::maybe_update() helper method accepted a mere
bool, indicating whether to convert the object being operated on into
the Message::Updated variant or leave it as-is. With an upcoming change
the Message::Updated variant will contain the ID(s) of the widget(s)
that need to be updated, rendering a bool argument to this helper method
insufficient.
With this change we adjusts MessageExt::maybe_update() to accept an
Option<Message> instead. In so doing we open the door for performing
more work in this method, including the merge of multiple IDs to update,
for example.