Introduce Ids type for representing a set of ID objects
We are working our way towards a world where the UI updates
incrementally, only redrawing widgets that indicated that they changed.
To make that vision a reality, we need a way to convey which widgets
need to be updated from the event handling path.
With this change we introduce the Ids type, which effectively is a set
of ID objects representing the widgets to update. Two objects of this
type can be merged, in which case the resulting object is the union of
the IDs captured by both objects.