Extend {Message,Event}::Updated variants with Ids
As motivated earlier, in order to implement incremental UI updates we
are going to need to convey the IDs of widgets that require a redraw
from the event/message handling path.
To make that happen, this change extends the Message::Update and
Event::Updated variants to hold an Ids object, which holds the ID
objects representing the set of widgets to update.