Fix cursor positioning in TaskListBox
With our recent refactoring of the TaskListBox rendering code, we
introduced a bug manifesting in a wrong cursor position on the selected
task: instead of the start of the task summary, it is the indication
whether or not a description is present that is where the cursor is
positioned on.
This change fixes the issue by adjusting the cursor position
accordingly.