Refine rendering of InOut widget
Similar to the TaskListBox (which was fixed earlier), the InOut widget
has some problems when it comes to rendering its contents in a scheme
where not the entire UI is rendered unconditionally: it does not fill
all parts of the widget but only those that contain "actual" data. While
that was fine so far, it will bite us once we switch over to rendering
parts of the UI incrementally.
To fix this issue, this change adjusts the InOut rendering code to
unconditionally fill any previously unused space with the corresponding
background color.