Introduce infrastructure for redrawing UI incrementally
With a follow-on change we will start redrawing the UI incrementally. In
order to get there we need mechanisms for
1) inferring the widgets to redraw
2) only redrawing these widgets

This change puts both in place. Specifically, we collect the IDs of all
widgets that require an update since the most recent redrawing. We then
communicate these IDs to the TermRenderer. When said renderer now
renders a widget, it checks whether this widget is in the set of widgets
to update. If so it remembers that fact. Similarly, whenever we are done
rendering a widget, we check whether we just finished the widget we
earlier flagged as requiring an update and, if so, reset the
corresponding flag again. Using this algorithm we can correctly track
the redrawing of a widget and its children, using only the ID of the
widget.