Redraw terminal UI incrementally
So far we have re-rendered the entire UI whenever a change occurred.
While not necessary inherent to this approach, the way we went about
doing it causes a bunch of rendering related artifacts, such as
flickering. With this change we switch over to an incremental rendering
approach, where we only render widgets (and their children) that
indicated that they have changed. This approach now relies on the
correct reporting of all changed widgets from the event and message
handling paths, as well as the fact that all widgets correctly redraw
their entire surface area.
We make use of the mechanisms introduced to the TermRenderer earlier,
which allow us to determine whether a widget actually needs to be
rendered. If that is not the case, we just skip the rendering step and
move on to the next widget.