Fix cursor positioning after InOut widget state change
When the InOut changes state and gets hidden as a result, specifically,
we need to make sure to redraw its parent widget. That is necessary,
because said parent (or one of its children) may want to set the cursor
position now that the InOut widget is no longer active.
To that end, this change adjusts the InOut widget's change_state()
method to emit a message indicating an update of the parent, not the
InOut widget itself.