Eliminate unnecessary updates for InputText
When readline support is enabled, the InputText type never return the
InputResult::Unchanged variant. That is because we never checked whether
something actually changed, we just assumed that it did.
With this change we address this short coming by checking whether either
the cursor or the text changed in response to input, and if that is not
the case then we indicate that by returning InputResult::Unchanged to
the caller. In so doing we eliminate a bunch of unnecessary redraws, for
example when the cursor is already in the left most position and the
user asks to have it moved further left.