Adjust Tasks::add() method to accept task Builder object
The Tasks::add() method is not particularly flexible, owing to the fact
that it effectively defines a set of attributes that make up a task
while ignoring others. Currently, for example, it is not possible to add
a task with non-empty details, because the method does not accept task
details. One can work around that by adding the task and then updating
it, but that results in two undo steps being created, which is not great
from a usability point of view.
This change adjusts the Tasks::add() method to accept a task Builder
object instead. By definition, such a Builder can capture all attributes
making up a Task, meaning that the aforementioned issues are
sidestepped somewhat elegantly.