Add another task copy & paste test
This change adds another test for the task copy & paste functionality.
This test makes sure that we can copy a task on one view and paste it on
another. If the task's set of tags does not match that of the current
view, the next one containing the task will be selected.