Build more combinations of statically linked binary
This change adjusts our 'Build' workflow to build not just the "regular"
version of the program but also one with readline support included. We
switch over from using musl libc to glibc, because we need to link
against the distribution provided libreadline.a archive, which includes
references to glibc itself.