Add proper CI job for building with minimum dependency versions
While we build the crate with a certain minimum version that we claim to
support, there is no guarantee that dependencies will not transparently
bump the Rust version they require. Because we do not check in our
Cargo.lock, CI will always work with the most recent versions of
dependencies.
In order to prevent any issues from this setup, this change introduces a
proper CI job that makes sure to compile against the minimum versions of
dependencies with the minimum Rust version we support.