Switch to using clap crate for argument handling matters
So far we have hand rolled out argument "parsing" logic. That was fine
given the small amount of (mostly mutually exclusive) arguments the
program had. However, with upcoming changes we intend to add more
arguments. As such, it makes sense to switch over to a "proper" argument
parser, as the conditions would explode and hand coding the logic would
be cumbersome and error-prone.