Factor out Paths type
Our internal path/directory management is a bit messy: we assume
different directories for tasks, configuration, and "volatile" state.
The main program always assumes tasks to be located in a tasks/
sub-folder inside the config directory, but tests actually do not.
Furthermore, directory storing volatile state doesn't really have to be
made configurable at all: conceptually it can just be inferred based on
the corresponding config directory.
This change reworks the core of this logic and moves it into the new
Paths type. This type nicely abstracts away lower level details, which
will allow us to make certain changes, such as deriving one path from
another, in the future. The only input that instantiation of this type
requires is a configuration directory -- and even that has a default
value.