Store "volatile" state relative to configuration
With the rework of directory and path management into the Paths type we
have already eliminated the need (and ability) for providing a custom
directory for storing "volatile" state. With this change we take things
one step further and store said state not in a single fixed location but
at a path that depends on the location of the configuration. By doing
that, we open the door for multiple instances of the program being
active at any point in time, as long as they work on different
configuration (and with that also task) directories.