Unpin unicode-width dependency
Commit 08c139f6ef41 ("Pin unicode-width to 0.1.11") pinned the
unicode-width dependency, because version 0.1.12 of the crate was
causing test failures due to changed behavior.
As it turned out, the unicode strings we use for testing contain
invisible bytes that screw us over, as they are reported as having a
width. To eliminate this source of confusion, replace the character
sequences with the correct glyph ('\u{26A0}' in Rust). The tests are
adjusted accordingly and the unicode-width dependency can be unpinned.