Fix potential rendering bugs
When the containing terminal is small -- smaller than our dialogs -- we
see overly long lines being rendered that end up being wrapped around by
the terminal, distorting overall output. The problem is that we are not
properly taking into account the terminal size when clipping string
before rendering them to the terminal.
With this change we address this shortcoming.