Enable line tables for CI Rust artifacts
With debug information completely disabled in CI, we lack potentially
important source code line information should a test panic and we see a
backtrace.
To make it easier to debug such problems when only a backtrace is
available, include line tables in the Rust artifacts so that we also see
source code file names and line numbers.