Automate update of UI configuration
As of commit cf022c00070b ("Split up UI configuration"), we effectively
support two configuration formats. However, while we were able to load
them both, we never really forced an automated conversion. As a result,
it was unlikely that a particular user's configuration was ever updated.
With this patch we modify the logic to detect if the old  view format is
present and, if so, force writing out the configuration on "save". As a
result, a format conversion will implicitly happen.