Introduce support for parsing logical formulas
So far our "views" were effectively serialized as structured sets of
positive or negative tags, making them largely unsuitable for entry by a
user, certainly from a UI.
In an attempt to eventually make these views configurable through the
UI, we want to support a more general notion of logical formulas, which
users can be trusted to craft themselves. As a first step towards this
vision, this change introduces the means for parsing such general
formulas from a textual representation. Currently support logical
operations are negation, logical AND, and logical OR. In the future we
should easily be able to extend this list. The parser is built around
the concept of parser combinators, which allow for a fairly high level
declarative description of the "language" in question based on lower
level building blocks.