Add conversion between Formula and CNF
In order to use the new Formula type, we need a way to convert such a
formula into Conjunctive Normal Form (CNF), which we use in our View.
Conversely, in order to transition our configuration away from storing
the CNF directly and towards using a serialized Formula, we need to be
able to convert from CNF to Formula.
With this change we introduce both of these conversions.