Only register event hook when truly necessary
The TermUi unconditionally registers an event hook and intercepts all
events, just for the sake of identifying the repeated press of the Q
key. That is pretty wasteful and unnecessary at a conceptual level.
Rework the logic to install the event hook only when necessary and
remove it once it has done its job.