Make ser::TagLit enum generic over tag type
Right now our ser::TagLit enum always works with actual Tag objects.
However, in the future we may want it to work more generally with
anything that could represent a tag more broadly.
To that end, this change makes this enum generic over the tag type to
use, while keeping the existing default.