Make Templates::instantiate_from_name() return an Option
So far our Templates::instantiate_from_name() helper was restricted
solely to test code. As such, it was fine for it to panic on invalid
input.
In the future we intend to use it in the program itself. As such, adjust
its signature to indicate instantiation failure by returning an Option
instead.