Represent tags in formulas as strings
So far the tags in a Formula were always assumed to be numbers. After
all, they represented the tag ID to instantiate. We eventually want
users to be able to edit these tag formulas via the UI. Dealing with IDs
in such a context is a no-go.
As such, with this change we switch our Formula type to represent a
"variable" as a string as opposed to a number. To make this work out
with respect to the serialization/deserialization logic, adjust it to
perform tag instantiation only as part of actual View creation and no
longer during serialization already. As a result, we will also make sure
to persist the new formula based view configuration when the user saves
changes.
Note that not all accepted template names are valid tags anymore at this
point, meaning some manual renaming of templates may be necessary to
eliminate the usage of special characters.