Add response_id to Input type
When the In/Out widget is asked to input some text, it ultimately send
the result back to the previously focused widget. That can be
problematic if the request originated with a widget that is more logical
in nature and, by-design, does not receive the input focus. Our TabBar
is such an example, as it basically just "manages" individual
TaskListBox objects, one of which will ultimately be focused. Input
handling requests from the TabBar are subtly broken in that
InputCanceled messages are never propagated to it and even EnteredText
handling takes a somewhat complicated detour through TaskListBox.

This change improves the situation by adding the ID of a widget to which
to send the response once input finished to the Input type.