Introduce and use KeyEvent alias
Introduce the KeyEvent type alias, representing a tuple of a key as well
as the raw key input data (in case the 'readline' feature is active) or
the empty tuple. Start using this type for backing the Event::Key
variant. In so doing we force all UI code to correctly plumb through the
full key event data, which will be required once we start re-emitting
events with upcoming changes.