Adjust Kseq widget to use input focus instead of event hooking
So far the Kseq widget has used the gui crate's event hooking mechanism
in order to inspect key sequences. The downside of this approach, as it
turned out, is that basically there is no way to stop the hooked event
from being received as well, meaning that even if a key sequence is
detected, the handler for the last key press may still be invoked. To
work around issues arising from this behavior, we previously introduced
the is_active() method to KseqData, but usage of it is extremely error
prone: a widget now needs to proactively decide to ignore an event when
Kseq is "active", which is easy to forget or do wrongly and has to
happen at all places, even when the widget itself doesn't use Kseq
itself and just overlaps with key bindings.
With this change we rework how Kseq works. Specifically, similar to the
InOut widget, we now make Kseq a "modal" and provide it with the input
focus while it captures a key sequence. Once it is done it will re-focus
the previously active widget. As before, it will send a GotKeySeq
message to the requesting widget as well, but to support behavior such
as the existing quit-quit workflow, we furthermore support re-emitting
the most recent key press to the newly focused widget. This way, we
enable client widgets to decide whether the last key press in a sequence
should be considered consumed or not.