Factor out generic key sequence recognition logic
In order to provide the quit-on-unsaved-changes logic we implement what
effectively amounts to a key sequence detection: one 'q' after another
'q' key press. However, the logic for detection and handling is very
much intertwined with the TermUi widget.
With this change we factor out a new widget type, Kseq, whose only
purpose is to detect key and report key sequences (we need a widget
because only widgets can hook events and we need to do that). In so
doing we make this functionality more easily reusable, which will come
in handy with upcoming changes.