Track Kseq widget "activity"
In the near future we would like to use the Kseq widget to register key
sequences that include key presses which already have a meaning in a
non-sequence context (think: only 'l' versus 'v' + 'l'). While possible,
because event hooks can't discard events, the "unsequenced" event will
still be delivered. Ideally, the gui crate would make event discarding a
support workflow, but for lack of that we can build our own "ignore
event" logic whenever we are actually in a key sequence.
This change lays the ground work for that, by tracking and exposing the
"activity" state of the Kseq widget. When the widget is active users can
just opt-in to ignore unrelated events.
Note that this change hinges on a small fix to the gui crate's post-hook
event delivery that got included with version 0.6.5 of the library that
allows us to keep the Kseq event hook installed until the post-hook is
actually invoked.