Eagerly de-focus Kseq widget on key sequence completion
When the Kseq widget receives a key press that completes a key sequence,
it informs some "receiver" widget about that fact. Only after that does
it restore the input focus to the previously focused widget.
Conceptually it makes more sense to restore the input focus before
sending out a message to the receiver, because the receiver may in turn
change input focus on its own, which would be overwritten once Kseq gets
control. Such behavior is never desirable.
To prevent it from happening, restore the focus to the previously
selected widget before sending out the message indicating key sequence
completion.