Simplify TaskListBox Message::InputCanceled handling
It is not clear why the TaskListBox widget asks for a rendering update
when receiving the InputCanceled message: the user's input is being
discarded and in both the task add and task edit cases no visible state
of the TaskListBox has been modified.
As such, simplify the code accordingly and do not ask for a re-render.